+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * Modified `create_EPOCH()` to accommodate the EPOCHST method using the `ST = TRUE` argument
 
+* Modified `convert_to_script()` so that the `filename` argument can handle no file extension or a .R or .Rmd file extension
+
 # sdtmval 0.4.0
 
 ## Major changes